chore: Update Settings page to set 'ANTHROPIC_API_KEY' value from user configs
The code changes include updating the Settings page to set the value of the 'ANTHROPIC_API_KEY' field in the form from the user's configurations. This ensures that the API key is pre-filled when the page loads. The eslint-disable-next-line comment has been added to ignore the exhaustive-deps warning.

Based on the recent user commits and repository commits, the commit message follows the established convention of starting with a type (feat, fix, chore) followed by a concise description of the changes made.