chore: Update package.json to include dayjs dependency
The code changes in the package.json file add the dayjs dependency with version 1.11.11. This is done to introduce the functionality provided by dayjs into the project.

Based on the recent user commits and repository commits, the commit message follows the established convention of starting with a type (feat, fix, chore) followed by a concise description of the changes made.